Rewrite ecompress to use per-file markings
Use per-file markings for every potentially compressible file rather
than a combination of per-file and per-directory 'compress' and 'skip'
markers. While this is less efficient, it is much simpler and more
reliable (and do we really expect to work on any data where this would
matter?). That said, make it a little more efficient by filtering
by file size straight away.

Since we no longer run compression per-directory, a lot of extra code
has been removed. As a part of that, the symlink fixing had to be
changed to apply to whole ED rather than explicitly compressed
directories only (which is an improvement since it doesn't leave broken
symlinks now).

Closes: https://github.com/mgorny/portage/issues/11