ecompress: Detect and report colliding (un)compressed files
Whenever the install directory contains files that would collide upon
(re)compressing, report them explicitly and fail.  This indicates
a serious problem in ebuild and since we don't know which of the files
is correct, we should not attempt to choose between them.

To reduce performance impact, the check is only done whenever compressed
files are found.  This is sufficient since for issue to occur there must
be at least one compressed variant.

Bug: https://bugs.gentoo.org/667072
Signed-off-by: Michał Górny <mgorny@gentoo.org>