Use verbose prefixes for output messages
Use verbose prefixes for output functions einfo, elog... in order to
make the message types distinguishable on non-color terminals,
and to enable the possibility of grepping them.  The prefixes are
roughly inspired by Xorg logs, and are:

- `[..]` for einfo (and ebegin)
- `[II]` for elog
- `[WW]` for ewarn
- `[QA]` for eqawarn
- `[EE]` for eerror

While at it, replace inline prefixes with references to a dict that
should make it easy to adjust them in the future.

Bug: https://bugs.gentoo.org/728046
Bug: https://bugs.gentoo.org/778860
Signed-off-by: Michał Górny <mgorny@gentoo.org>